backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
google-generativeai==0.8.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.3.1
openai==1.97.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
PyPDF2==3.0.1
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3

